+++ activesupport/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   Since weeks are no longer converted to days, add `:weeks` to the list of
+    parts that `ActiveSupport::TimeWithZone` will recognize as possibly being
+    of variable duration to take account of DST transitions.
+
+    Fixes #26039.
+
+    *Andrew White*
+
 *   Defines `Regexp.match?` for Ruby versions prior to 2.4. The predicate
     has the same interface, but it does not have the performance boost. Its
     purpose is to be able to write 2.4 compatible code.